% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_fabric.R
\name{make_fabric}
\alias{make_fabric}
\title{Make Reference Fabric}
\usage{
make_fabric(...)
}
\arguments{
\item{...}{Additional arguments passed to \link[targets:tar_make]{targets::tar_make}.}
}
\description{
Make Reference Fabric
}
